UI module update
Updated UI module to simplify the code.  Moved the rect and z variable
as well as visible variable to base UI Element class, as those three
were common to all classes.  I also updated the Engine to use a variable
for background color, and updated naming of a variable to be more
consistent.